+++ outputs.tf
@@ -147,12 +147,12 @@ output "cluster_tls_certificate_sha1_fingerprint" {
 ###############################################################################
 
 output "cluster_iam_role_name" {
-  description = "IAM role name of the EKS cluster"
+  description = "Cluster IAM role name"
   value       = try(module.eks[0].cluster_iam_role_name, null)
 }
 
 output "cluster_iam_role_arn" {
-  description = "IAM role ARN of the EKS cluster"
+  description = "Cluster IAM role ARN"
   value       = try(module.eks[0].cluster_iam_role_arn, null)
 }
 
@@ -160,6 +160,28 @@ output "cluster_iam_role_unique_id" {
   description = "Stable and unique string identifying the IAM role"
   value       = try(module.eks[0].cluster_iam_role_unique_id, null)
 }
+
+
+################################################################################
+# EKS Auto Node IAM Role
+################################################################################
+
+output "node_iam_role_name" {
+  description = "EKS Auto node IAM role name"
+  value       = try(module.eks[0].node_iam_role_name, null)
+}
+
+output "node_iam_role_arn" {
+  description = "EKS Auto node IAM role ARN"
+  value       = try(module.eks[0].node_iam_role_arn, null)
+}
+
+output "node_iam_role_unique_id" {
+  description = "Stable and unique string identifying the IAM role"
+  value       = try(module.eks[0].node_iam_role_unique_id, null)
+}
+
+########
 
 ################################################################################
 # EKS Addons