+++ variables_03_argo.tf
@@ -7,7 +7,7 @@ variable "argocd" {
     extra_command_arg  = optional(list(string), [])
     chart_repo         = optional(string, "https://argoproj.github.io/argo-helm")
     chart_name         = optional(string, "argo-cd")
-    chart_version      = optional(string, "7.1.3")
+    chart_version      = optional(string, "7.7.16")
     helm_release_name  = optional(string, "argocd")
     app_project_name   = optional(string, "cluster-addons")
     path_to_values     = optional(string, "argocd-values.yaml")
@@ -23,7 +23,7 @@ variable "argocd" {
       path              = optional(string, "")
       chart_repo        = optional(string, "https://argoproj.github.io/argo-helm")
       chart_name        = optional(string, "argocd-apps")
-      chart_version     = optional(string, "2.0.0")
+      chart_version     = optional(string, "2.0.2")
       helm_release_name = optional(string, "argocd-apps")
       repository = optional(object({
         path           = optional(string, "")