+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2025-01-18
+### Updated
+- Version EKS module
+- Version Karpenter module
+- fix for kubernetes_secret oidc_secret 
+
 ## [1.1.0] - 2024-10-20
 ### Updated
 - Version EKS module